+++ src/routes/router.tsx
@@ -1,3 +1,4 @@
+import GoogleCallbackPage from '@/components/pages/GoogleCallbackPage';
 import LoginPage from '@/components/pages/LoginPage';
 import MainPage from '@/components/pages/MainPage';
 import RolePage from '@/components/pages/RolePage';
@@ -10,6 +11,10 @@ const router = [
   {
     path: '/roleselect',
     element: <RolePage />,
+  },
+  {
+    path: '/auth/google/callback',
+    element: <GoogleCallbackPage />,
   },
   {
     path: '/',
